feat(auth): add Clerk authentication to [POST] /api/v1/git_tokens/
Summary:
- updated API '[POST] /api/v1/git_tokens/' to start using Clerk authentication.

- Exception was being swallowed and not logged, added logging.

- Fixed Unit Tests related to this API

- Moved the unit test to a proper structure
