+++ tests/test_token_api.py
@@ -7,7 +7,6 @@ from fastapi import status
 import pytest
 from urllib.parse import quote
 from unittest.mock import patch, MagicMock, AsyncMock
-from app.routes.git_tokens import mask_token
 from app.utils.auth import AuthenticatedUserDTO, get_current_user
 from app.main import app
 
@@ -177,176 +176,6 @@ class TestGetGitLabelsEndpoint:
 
             mock_query.offset.assert_called_with(2)  # (page-1) * size = (2-1) * 20 = 20
             mock_query.limit.assert_called_with(20)
-
-
-
-
-class TestAddGitTokenEndpoint:
-    """Test cases for POST /api/v1/git_tokens/ endpoint."""
-
-    def test_add_github_token_success(
-        self, client, token_payload_github, mock_encryption_helper
-    ):
-        """Test successful creation of a GitHub token."""
-        with patch("app.routes.git_tokens.GitHubManager") as mock_github_manager:
-            # Setup successful GitHub manager
-            mock_instance = MagicMock()
-            mock_instance.get_user.return_value = {"login": "testuser"}
-            mock_github_manager.return_value = mock_instance
-
-            with patch("app.routes.git_tokens.GitLabel") as mock_git_label:
-                # Setup successful GitLabel creation
-                mock_created_label = MagicMock()
-                mock_created_label.id = "999"
-                mock_git_label.create = AsyncMock(return_value=mock_created_label)
-
-                response = client.post("/api/v1/git_tokens/", json=token_payload_github)
-
-                assert response.status_code == status.HTTP_200_OK
-                data = response.json()
-                assert data["success"] is True
-                assert data["data"]["id"] == "999"
-
-                # Verify GitLabel.create was called
-                mock_git_label.create.assert_called_once()
-
-    def test_add_gitlab_token_success(
-        self, client, token_payload_gitlab, mock_encryption_helper
-    ):
-        """Test successful creation of a GitLab token."""
-        with patch("app.routes.git_tokens.GitLabManager") as mock_gitlab_manager:
-            # Setup successful GitLab manager
-            mock_instance = MagicMock()
-            mock_instance.auth_status = True
-            mock_instance.get_user.return_value = {"username": "testuser"}
-            mock_gitlab_manager.return_value = mock_instance
-
-            with patch("app.routes.git_tokens.GitLabel") as mock_git_label:
-                mock_created_label = MagicMock()
-                mock_created_label.id = "999"
-                mock_git_label.create = AsyncMock(return_value=mock_created_label)
-
-                response = client.post("/api/v1/git_tokens/", json=token_payload_gitlab)
-
-                assert response.status_code == status.HTTP_200_OK
-                data = response.json()
-                assert data["success"] is True
-                assert data["data"]["id"] == "999"
-
-    def test_add_github_token_authentication_failure(
-        self, client, token_payload_github, mock_encryption_helper
-    ):
-        """Test GitHub token creation with authentication failure."""
-        with patch("app.routes.git_tokens.GitHubManager") as mock_github_manager:
-            mock_instance = MagicMock()
-            mock_instance.get_user.return_value = None
-            mock_github_manager.return_value = mock_instance
-
-            response = client.post("/api/v1/git_tokens/", json=token_payload_github)
-
-            assert response.status_code == status.HTTP_400_BAD_REQUEST
-            data = response.json()
-            assert data["success"] is False
-            assert "Failed to authenticate with GitHub" in data["message"]
-
-    def test_add_gitlab_token_authentication_failure(
-        self, client, token_payload_gitlab, mock_encryption_helper
-    ):
-        """Test GitLab token creation with authentication failure."""
-        with patch("app.routes.git_tokens.GitLabManager") as mock_gitlab_manager:
-            # Setup failed GitLab manager
-            mock_instance = MagicMock()
-            mock_instance.auth_status = False
-            mock_gitlab_manager.return_value = mock_instance
-
-            response = client.post("/api/v1/git_tokens/", json=token_payload_gitlab)
-
-            assert response.status_code == status.HTTP_400_BAD_REQUEST
-            data = response.json()
-            assert data["success"] is False
-            assert "Failed to authenticate with GitLab" in data["message"]
-
-    def test_add_gitlab_token_user_fetch_failure(self, client, mock_encryption_helper):
-        """Test GitLab token creation with user fetch failure."""
-        with patch("app.routes.git_tokens.GitLabManager") as mock_gitlab_manager:
-            # Mock GitLab auth success but user fetch failure
-            mock_instance = MagicMock()
-            mock_instance.auth_status = True
-            mock_instance.get_user.return_value = None
-            mock_gitlab_manager.return_value = mock_instance
-
-            payload = {
-                "label": "Test GitLab Token",
-                "git_hosting": "gitlab",
-                "token_value": "glpat-1234567890abcdef",
-            }
-
-            response = client.post("/api/v1/git_tokens/", json=payload)
-
-            assert response.status_code == status.HTTP_400_BAD_REQUEST
-            data = response.json()
-            assert data["success"] is False
-            assert "Could not retrieve GitLab user" in data["message"]
-
-            # Verify user data was attempted to be fetched
-            mock_gitlab_manager.return_value.get_user.assert_called_once()
-
-    def test_add_token_unsupported_provider(self, client, mock_encryption_helper):
-        """Test creation with unsupported git hosting provider."""
-        payload = {
-            "label": "Test Token",
-            "git_hosting": "bitbucket",  # Unsupported provider
-            "token_value": "token123456",
-        }
-
-        response = client.post("/api/v1/git_tokens/", json=payload)
-
-        assert response.status_code == status.HTTP_422_UNPROCESSABLE_ENTITY
-        data = response.json()
-        assert data["success"] is False
-        assert "Unsupported git hosting provider" in data["message"]
-
-    def test_add_token_empty_token(self, client, mock_encryption_helper):
-        """Test creation with empty token."""
-        payload = {
-            "label": "Test GitHub Token",
-            "git_hosting": "github",
-            "token_value": "",  # Empty token
-        }
-
-        response = client.post("/api/v1/git_tokens/", json=payload)
-
-        assert response.status_code in [
-            status.HTTP_422_UNPROCESSABLE_ENTITY,
-            status.HTTP_400_BAD_REQUEST,
-        ]
-
-    def test_user_id_override_security(self, client, mock_encryption_helper):
-        """Test that user_id is overridden with authenticated user for security."""
-        with patch("app.routes.git_tokens.GitHubManager") as mock_github_manager:
-            mock_instance = MagicMock()
-            mock_instance.get_user.return_value = {"login": "testuser"}
-            mock_github_manager.return_value = mock_instance
-
-            with patch("app.routes.git_tokens.GitLabel") as mock_git_label:
-                mock_created_label = MagicMock()
-                mock_created_label.id = "999"
-                mock_git_label.create = AsyncMock(return_value=mock_created_label)
-
-                payload = {
-                    "label": "Test GitHub Token",
-                    "git_hosting": "github",
-                    "token_value": "ghp_1234567890abcdef",
-                    "user_id": "malicious-user-id",  # This should be overridden
-                }
-
-                response = client.post("/api/v1/git_tokens/", json=payload)
-
-                assert response.status_code == status.HTTP_200_OK
-
-                # Verify that GitLabel.create was called with the authenticated user ID, not the provided one
-                call_args = mock_git_label.create.call_args
-                assert call_args.kwargs["user_id"] == "user-123"
 
 
 class TestDeleteGitLabelEndpoint:
